requirements.txt
# Core Deep Learning
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Scientific Computing
scipy>=1.11.0
numpy>=1.24.0

# Configuration and Utilities
pyyaml>=6.0
tqdm>=4.66.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For better logging
tensorboard>=2.14.0
